# api/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: longue-vue CMDB API
  version: 0.1.0
  summary: Kubernetes-centric CMDB aligned with the ANSSI SecNumCloud framework.
  description: |
    REST API for **longue-vue**, a CMDB for Kubernetes environments aligned with the
    ANSSI SecNumCloud (SNC) qualification framework.

    This version (0.1.0) is a walking skeleton covering:
      - Liveness and readiness probes
      - `Cluster` resource CRUD

    Subsequent versions will add Node, Namespace, Workload, Pod, and other
    Kubernetes entity kinds mapped onto the ANSSI cartography layers.

    Errors follow RFC 7807 (`application/problem+json`).

    Interactive API documentation is available at `/docs/` for authenticated
    operators; sign in via `/ui/login` first to use "Try it out".
servers:
  - url: /
    description: Current origin (recommended for browser-based "Try it out" via /docs/)
  - url: http://localhost:8080
    description: Local development
...
security:
  - BearerAuth: []
  - SessionCookie: []
paths:
...
  /v1/applications:
    get:
      tags: [applications]
      summary: List applications
      operationId: listApplications
      security:
        - BearerAuth: [read]
        - SessionCookie: [read]
      parameters:
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
        - name: name
          in: query
          required: false
          description: Case-insensitive substring filter on name + display_name.
          schema: { type: string }
        - name: application_block_id
          in: query
          required: false
          description: Restrict to applications belonging to this block.
          schema:
            type: string
            format: uuid
        - name: application_block_name
          in: query
          required: false
          description: Restrict to applications belonging to the block with this exact name.
          schema: { type: string }
        - name: criticality
          in: query
          required: false
          description: Exact criticality filter.
          schema: { type: string }
        - name: has_dict
          in: query
          required: false
          description: |
            When true, restrict to applications with at least one DICT
            classification field set. When false, restrict to applications
            with none set.
          schema: { type: boolean }
        - name: dict_min
          in: query
          required: false
          description: |
            Restrict to applications whose `MAX(sec_*)` is at least this
            value (SNC severity floor filter).
          schema:
            type: integer
            minimum: 0
            maximum: 4
      responses:
        '200':
          description: Paged list of applications.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApplicationList'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'

    post:
      tags: [applications]
      summary: Create an application
      description: |
        Either `application_block_id` (UUID) or `application_block_name`
        (string) may be supplied to link the new application to a block.
        If both are present the UUID wins.
      operationId: createApplication
      security:
        - BearerAuth: [write]
        - SessionCookie: [write]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ApplicationCreate'
      responses:
        '201':
          description: Application created.
          headers:
            Location:
              description: URI of the newly created application.
              schema:
                type: string
                format: uri
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Application'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: token
      description: |
        API token issued out-of-band. Send as `Authorization: Bearer <token>`.

        Scopes granted to a token:
          - `read`   — list and get operations
          - `write`  — create and update (implies `read` at the caller's discretion)
          - `delete` — remove resources
          - `admin`  — all of the above; reserved for future admin-only endpoints

        Tokens carrying the `admin` scope implicitly satisfy any scope required
        by an operation.

        Per ADR-0007, machine tokens are created by an admin in the UI
        (see `POST /v1/admin/tokens`) rather than injected through env vars.

    SessionCookie:
      type: apiKey
      in: cookie
      name: longue_vue_session
      description: |
        Human session cookie set by `POST /v1/auth/login`. HttpOnly,
        SameSite=Strict, 8-hour sliding expiry. Per ADR-0007.

        Every operation accepts either `BearerAuth` or `SessionCookie` —
        scope enforcement reads from the caller's role (for sessions) or
        token scopes (for bearer), so the two paths share the same
        downstream authorisation check.
  parameters:
...
    Limit:
      name: limit
      in: query
      required: false
      description: Maximum number of items to return. Server clamps to [1, 200].
      schema:
        type: integer
        minimum: 1
        maximum: 200
        default: 50
    Cursor:
      name: cursor
      in: query
      required: false
      description: Opaque cursor returned from a previous list response.
      schema:
        type: string
        maxLength: 512
...
  responses:
    BadRequest:
      description: Malformed request.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    Unauthorized:
      description: Missing or invalid bearer token.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    Forbidden:
      description: Token is valid but lacks the scope required by the operation.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    NotFound:
      description: Resource not found.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    Conflict:
      description: Conflict with current state (e.g., duplicate name).
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
  schemas:
...
    Problem:
      description: RFC 7807 problem details.
      type: object
      required: [type, title, status]
      properties:
        type:
          type: string
          format: uri
          default: about:blank
        title:
          type: string
        status:
          type: integer
          minimum: 100
          maximum: 599
        detail:
          type: string
        instance:
          type: string
          format: uri
        errors:
          type: array
          description: Field-level validation errors, when applicable.
          items:
            type: object
            required: [field, message]
            properties:
              field:
                type: string
              message:
                type: string
...
    ApplicationMemberCounts:
      type: object
      required: [workloads, virtual_machines, vm_applications]
      properties:
        workloads:        { type: integer }
        virtual_machines: { type: integer }
        vm_applications:  { type: integer }
    Application:
      type: object
      required: [id, name, annotations, created_at, updated_at, member_counts]
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        display_name: { type: string, nullable: true }
        description: { type: string, nullable: true }
        application_block_id: { type: string, format: uuid, nullable: true }
        application_block_name: { type: string, nullable: true, readOnly: true }
        owner: { type: string, nullable: true }
        criticality: { type: string, nullable: true }
        notes: { type: string, nullable: true }
        runbook_url: { type: string, nullable: true }
        annotations:
          type: object
          additionalProperties: { type: string }
        sec_disponibilite:   { type: integer, minimum: 0, maximum: 4, nullable: true }
        sec_integrite:       { type: integer, minimum: 0, maximum: 4, nullable: true }
        sec_confidentialite: { type: integer, minimum: 0, maximum: 4, nullable: true }
        sec_tracabilite:     { type: integer, minimum: 0, maximum: 4, nullable: true }
        sec_notes: { type: string, nullable: true }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
        member_counts: { $ref: '#/components/schemas/ApplicationMemberCounts' }
    ApplicationCreate:
      type: object
      required: [name]
      properties:
        name: { type: string }
        display_name: { type: string }
        description: { type: string }
        application_block_id: { type: string, format: uuid }
        application_block_name: { type: string }
        owner: { type: string }
        criticality: { type: string }
        notes: { type: string }
        runbook_url: { type: string }
        annotations: { type: object, additionalProperties: { type: string } }
        sec_disponibilite:   { type: integer, minimum: 0, maximum: 4 }
        sec_integrite:       { type: integer, minimum: 0, maximum: 4 }
        sec_confidentialite: { type: integer, minimum: 0, maximum: 4 }
        sec_tracabilite:     { type: integer, minimum: 0, maximum: 4 }
        sec_notes: { type: string }
...
    ApplicationList:
      description: Paged list of applications.
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Application'
        next_cursor:
          type: string
          nullable: true
          description: |
            Opaque cursor to pass as `?cursor=` to fetch the next page.
            Absent or null when no more pages remain.
...